Removing unnecessary tables
We had a set of old tables that were used for cog/over assignment. Those
tables are no longer in use.

Therefore, we no longer need to make sure they are part of our backup.

This commit removes those tables from the manifest, so that we stop
asking if they are present.